+++ packages/apollo-boost/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change log
 
+### vNext
+- Map coverage to original source
+
 ### 0.1.0
 - DEPRECATED: `apollo-client-preset`
 - Changed to `apollo-boost` [#2965](https://github.com/apollographql/apollo-client/pull/2965)