+++ packages/apollo-boost/CHANGELOG.md
@@ -1,8 +1,14 @@
 # Change log
 
-### vNEXT
+### vNext
+- Map coverage to original source
+- fix request parameter type [#3056](https://github.com/apollographql/apollo-client/pull/3056)
 
-### 1.0.9
+### 0.1.0
+- DEPRECATED: `apollo-client-preset`
+- Changed to `apollo-boost` [#2965](https://github.com/apollographql/apollo-client/pull/2965)
+
+### 1.0.9 (last `apollo-client-preset` version)
 - Apollo Client 2.2.3
 
 ### 1.0.8